Shared: Move numbers helpers to shared sub repo (#932)
* chore: move numbers helpers to shared repo

* voting: fix numbers helper import
