+++ apps/voting/test/contracts/voting.js
@@ -1,8 +1,8 @@
 const VOTER_STATE = require('../helpers/state')
 const getBlockNumber = require('@aragon/test-helpers/blockNumber')(web3)
-const { getVoteState } = require('../helpers/voting')
-const { bigExp, pct } = require('../helpers/numbers')(web3)
+const { bigExp } = require('@aragon/test-helpers/numbers')(web3)
 const { assertRevert } = require('@aragon/test-helpers/assertThrow')
+const { pct, getVoteState } = require('../helpers/voting')(web3)
 const { makeErrorMappingProxy } = require('@aragon/test-helpers/utils')
 const { encodeCallScript, EMPTY_SCRIPT } = require('@aragon/test-helpers/evmScript')
 const { assertEvent, assertAmountOfEvents } = require('@aragon/test-helpers/assertEvent')(web3)
