+++ apps/voting/test/contracts/voting_delegation.js
@@ -1,9 +1,9 @@
 const VOTER_STATE = require('../helpers/state')
-const { getVoteState } = require('../helpers/voting')
-const { bigExp, pct } = require('../helpers/numbers')(web3)
+const { bigExp } = require('@aragon/test-helpers/numbers')(web3)
 const { assertRevert } = require('@aragon/test-helpers/assertThrow')
 const { skipCoverage } = require('@aragon/os/test/helpers/coverage')
 const { encodeCallScript } = require('@aragon/test-helpers/evmScript')
+const { pct, getVoteState } = require('../helpers/voting')(web3)
 const { decodeEventsOfType } = require('@aragon/os/test/helpers/decodeEvent')
 const { makeErrorMappingProxy } = require('@aragon/test-helpers/utils')
 const { getEventArgument, getNewProxyAddress } = require('@aragon/test-helpers/events')
