+++ future-apps/payroll/test/contracts/Payroll_employee_info.test.js
@@ -1,7 +1,7 @@
 const { assertRevert } = require('@aragon/test-helpers/assertThrow')
 const { getEventArgument } = require('@aragon/test-helpers/events')
-const { annualSalaryPerSecond } = require('../helpers/numbers')(web3)
-const { MAX_UINT256, MAX_UINT64 } = require('../helpers/numbers')(web3)
+const { annualSalaryPerSecond } = require('../helpers/salary')(web3)
+const { MAX_UINT256, MAX_UINT64 } = require('@aragon/test-helpers/numbers')(web3)
 const { NOW, ONE_MONTH, RATE_EXPIRATION_TIME } = require('../helpers/time')
 const { deployContracts, createPayrollAndPriceFeed } = require('../helpers/deploy')(artifacts, web3)
 const { USD, deployANT, deployDAI } = require('../helpers/tokens')(artifacts, web3)
