+++ future-apps/payroll/test/contracts/Payroll_terminate_employee.test.js
@@ -1,9 +1,10 @@
 const PAYMENT_TYPES = require('../helpers/payment_types')
 const { assertRevert } = require('@aragon/test-helpers/assertThrow')
+const { annualSalaryPerSecond } = require('../helpers/salary')(web3)
+const { bn, bigExp, MAX_UINT64 } = require('@aragon/test-helpers/numbers')(web3)
 const { getEvents, getEventArgument } = require('@aragon/test-helpers/events')
 const { NOW, ONE_MONTH, RATE_EXPIRATION_TIME } = require('../helpers/time')
 const { deployContracts, createPayrollAndPriceFeed } = require('../helpers/deploy')(artifacts, web3)
-const { bn, bigExp, MAX_UINT64, annualSalaryPerSecond } = require('../helpers/numbers')(web3)
 const { USD, DAI_RATE, exchangedAmount, inverseRate, deployDAI, setTokenRate } = require('../helpers/tokens')(artifacts, web3)
 
 contract('Payroll employees termination', ([owner, employee, anyone]) => {
